fix: update error message and test cases
