+++ x509/codesigning_cert_validations.go
@@ -41,7 +41,7 @@ func ValidateCodeSigningCertChain(certChain []*x509.Certificate, signingTime *ti
 		}
 		// check self-issued
 		if !bytes.Equal(cert.RawSubject, cert.RawIssuer) {
-			return fmt.Errorf("invalid self-signed certificate. subject: %q. Error: issuer and subject are not the same", cert.Subject)
+			return fmt.Errorf("invalid self-signed certificate. subject: %q. Error: issuer(%s) and subject(%s) are not the same", cert.Subject, cert.Issuer, cert.Subject)
 		}
 		if signedTimeError := validateSigningTime(cert, signingTime); signedTimeError != nil {
 			return signedTimeError
